# refactor: Optimize getUserIdMappingWithEitherSuperTokensUserIdOrExternalUserId query